library/pics: fix parsing, DB saving and implement playtime fetcher

+++ core/src/main/kotlin/bruhcollective/itaysonlab/ksteam/models/library/LibraryCollection.kt
@@ -11,7 +11,7 @@ data class LibraryCollection(
     val removed: List<AppId>,
     val filterSpec: DynamicFilterSpec?,
     internal val timestamp: Int,
-    internal val version: Int,
+    internal val version: Long,
 ) {
     fun toRemoteModel() = CollectionModel(id, name, added.map(AppId::id), removed.map(AppId::id), filterSpec)
